kuwala-pipelines/google-poi/requirements.txt
aiofiles==0.6.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
priority==1.3.0
PySocks==1.7.1
python-dotenv==0.17.1
Quart==0.14.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wsproto==1.0.0
